% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_infos_stations.R
\name{get_infos_stations}
\alias{get_infos_stations}
\title{Informations sur une station Météo France}
\usage{
get_infos_stations(id_station, token)
}
\arguments{
\item{id_station}{Identifiant de la station (chaîne de caractères).}

\item{token}{Jeton d'accès à l'API Météo France.}
}
\value{
Un data.frame contenant les informations sur la station.
}
\description{
Renvoie l'historique de la station Météo France.
}
\examples{
if (interactive()) { 
token <- Sys.getenv("MF_TOKEN") 
infos_station <- get_infos_stations("35281001", token) 
print(infos_station)
 }
}
