% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mf_token.R
\name{get_mf_token}
\alias{get_mf_token}
\title{Récupérer le token Météo France}
\usage{
get_mf_token()
}
\value{
Une chaîne de caractères contenant le token.
}
\description{
Cette fonction lit la variable d'environnement \code{MF_TOKEN} et retourne sa valeur.
}
\examples{
if (interactive()) {
  Sys.setenv(MF_TOKEN = "fake_token")
  token <- get_mf_token()
  print(token)
}
}
