% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_precip_stations.R
\name{get_precip_stations}
\alias{get_precip_stations}
\title{get_precip_stations}
\usage{
get_precip_stations(departements, token)
}
\arguments{
\item{departements}{Vecteur avec la liste des codes de départements}

\item{token}{: token d'accès à l'API météo France (disponible gratuitement sur le site https://portail-api.meteofrance.fr/)}
}
\value{
Data Frame avec les stations météo disponibles
}
\description{
get_precip_stations
}
\examples{
if (interactive()) {
  token <- Sys.getenv("MF_TOKEN")  # ou demander à l'utilisateur
  stations <- get_precip_stations(c(56, 35), token)
  print(stations)
}
}
