% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametres_meteo.R
\docType{data}
\name{parametres_meteo}
\alias{parametres_meteo}
\title{Paramètres climatiques horaires}
\format{
Un data frame avec les colonnes :
\describe{
\item{mnemonique}{Code du paramètre}
\item{libelle}{Libellé du paramètre}
\item{unite}{Unité de mesure}
\item{pas_de_temps}{Pas de temps (ex: horaire)}
}
}
\source{
Données issues de Météo-France
}
\usage{
parametres_meteo
}
\description{
Jeu de données contenant les paramètres horaires disponibles via l'API Météo-France.
}
\keyword{datasets}
