% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telecharger_commande_csv.R
\name{telecharger_commande_csv}
\alias{telecharger_commande_csv}
\title{Télécharger un fichier CSV depuis une commande Météo France}
\usage{
telecharger_commande_csv(id_cmde, token)
}
\arguments{
\item{id_cmde}{Identifiant de la commande (chaîne de caractères).}

\item{token}{Jeton d'accès à l'API Météo France.}
}
\value{
Un data.frame contenant les données du fichier CSV.
}
\description{
Télécharge et lit un fichier CSV à partir d'un identifiant de commande.
}
\details{
Correspondance des colonnes principales :
\itemize{
\item POSTE : Identifiant de la station météo
\item DATE : Date et heure de l’observation (format AMJH)
\item RR1 : Précipitations horaires (mm)
\item QRR1 : Qualité de la donnée RR1
\item T : Température instantanée (°C)
\item QT : Qualité de la température
\item TN : Température minimale horaire
\item TX : Température maximale horaire
\item FF : Vitesse moyenne du vent
\item QFF : Qualité de la vitesse du vent
\item DD : Direction du vent (degrés)
\item FXI : Rafale instantanée maximale
\item PMER : Pression au niveau de la mer (hPa)
\item UV_INDICE : Indice UV
\item N : Nébulosité totale
\item CL : Nébulosité basse
\item CM : Nébulosité moyenne
\item CH : Nébulosité haute
}
}
\examples{
if (interactive()) { 
token <- Sys.getenv("MF_TOKEN") 
df <- telecharger_commande_csv("2025006008357", token) 
head(df) 
 }
}
